kok_bulma
This Go program calculates the roots of a quadratic equation (\( ax^2 + bx + c = 0 \)) using the coefficients **a, b, c** provided by the user. The **delta (\( \Delta = b^2 - 4ac \))** value is determined, and the roots of the equation are found using the discriminant method and printed to the screen. However, **the case of a negative delta is not handled**, meaning no processing is done for complex roots.